refactor: IO-422 create folder for report footers
