+++ src/components/Report/PdfReports/reportFooters/StrategyReportFooter.tsx
